Fix import merge, move ImportAdder to separate file
