+++ src/main/kotlin/org/elm/ide/intentions/MakeDecoderIntention.kt
@@ -1,5 +1,6 @@
 package org.elm.ide.intentions
 
+import org.elm.lang.core.imports.ImportAdder.Candidate
 import org.elm.lang.core.lookup.ElmLookup
 import org.elm.lang.core.psi.ElmFile
 import org.elm.lang.core.psi.elements.ElmTypeDeclaration
